Add possibility to show list of abandoned colonies (#1087)
* Add possibility to show list of abandoned colonies

since hours

* deleted not needed command

* fix review

* fix minor issues

* change to automatic detection

* make local var final
