+++ src/main/java/com/minecolonies/coremod/commands/citizencommands/AbstractCitizensCommands.java
@@ -1,8 +1,9 @@
-package com.minecolonies.coremod.commands;
+package com.minecolonies.coremod.commands.citizencommands;
 
 import com.minecolonies.coremod.colony.Colony;
 import com.minecolonies.coremod.colony.ColonyManager;
 import com.minecolonies.coremod.colony.IColony;
+import com.minecolonies.coremod.commands.AbstractSingleCommand;
 import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommandSender;
 import net.minecraft.entity.player.EntityPlayer;
