+++ src/main/java/com/minecolonies/coremod/commands/citizencommands/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains all commands related to citizen.
+ */
+package com.minecolonies.coremod.commands.citizencommands;
