+++ src/main/java/com/minecolonies/coremod/commands/colonycommands/ListColoniesCommand.java
@@ -1,7 +1,8 @@
-package com.minecolonies.coremod.commands;
+package com.minecolonies.coremod.commands.colonycommands;
 
 import com.minecolonies.coremod.colony.Colony;
 import com.minecolonies.coremod.colony.ColonyManager;
+import com.minecolonies.coremod.commands.AbstractSingleCommand;
 import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommandSender;
 import net.minecraft.server.MinecraftServer;
@@ -58,25 +59,26 @@ public class ListColoniesCommand extends AbstractSingleCommand
     @Override
     public void execute(@NotNull final MinecraftServer server, @NotNull final ICommandSender sender, @NotNull final String... args) throws CommandException
     {
-        int page = 1;
-        final List<Colony> colonies = ColonyManager.getColonies();
+        int page = getIthArgument(args, 0, 1);
+        final int abandonedSince = getIthArgument(args, 1, 0);
+
+        final List<Colony> colonies;
+        if (abandonedSince > 0)
+        {
+            colonies = ColonyManager.getColoniesAbandonedSince(abandonedSince);
+        }
+        else
+        {
+            colonies = ColonyManager.getColonies();
+        }
+
         final int colonyCount = colonies.size();
 
         // check to see if we have to add one page to show the half page
         final int halfPage = (colonyCount % COLONIES_ON_PAGE == 0) ? 0 : 1;
         final int pageCount = ((colonyCount) / COLONIES_ON_PAGE) + halfPage;
 
-        if (args.length != 0)
-        {
-            try
-            {
-                page = Integer.parseInt(args[0]);
-            }
-            catch (final NumberFormatException e)
-            {
-                //ignore and keep page 1.
-            }
-        }
+
         if (page < 1 || page > pageCount)
         {
             page = 1;
@@ -104,24 +106,23 @@ public class ListColoniesCommand extends AbstractSingleCommand
         for (final Colony colony : coloniesPage)
         {
             sender.addChatMessage(new TextComponentString(String.format(
-              ID_AND_NAME_TEXT, colony.getID(), colony.getName())).setStyle(new Style().setClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND,
-                                                                                                                           String.format(COMMAND_COLONY_INFO, colony.getID())))));
+                    ID_AND_NAME_TEXT, colony.getID(), colony.getName())).setStyle(new Style().setClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND,
+                    String.format(COMMAND_COLONY_INFO, colony.getID())))));
             final BlockPos center = colony.getCenter();
 
             final ITextComponent teleport = new TextComponentString(COORDINATES_TEXT + String.format(COORDINATES_XYZ, center.getX(), center.getY(), center.getZ()))
                     .setStyle(new Style().setBold(true).setColor(TextFormatting.GOLD).setClickEvent(
-                    new ClickEvent(ClickEvent.Action.RUN_COMMAND, TELEPORT_COMMAND + colony.getID())
-            ));
+                            new ClickEvent(ClickEvent.Action.RUN_COMMAND, TELEPORT_COMMAND + colony.getID())
+                    ));
 
             sender.addChatMessage(teleport);
-
         }
 
         final ITextComponent prevButton = new TextComponentString(PREV_PAGE).setStyle(new Style().setBold(true).setColor(TextFormatting.GOLD).setClickEvent(
-          new ClickEvent(ClickEvent.Action.RUN_COMMAND, LIST_COMMAND_SUGGESTED + prevPage)
+                new ClickEvent(ClickEvent.Action.RUN_COMMAND, LIST_COMMAND_SUGGESTED + prevPage)
         ));
         final ITextComponent nextButton = new TextComponentString(NEXT_PAGE).setStyle(new Style().setBold(true).setColor(TextFormatting.GOLD).setClickEvent(
-          new ClickEvent(ClickEvent.Action.RUN_COMMAND, LIST_COMMAND_SUGGESTED + nextPage)
+                new ClickEvent(ClickEvent.Action.RUN_COMMAND, LIST_COMMAND_SUGGESTED + nextPage)
         ));
 
         final ITextComponent beginLine = new TextComponentString(PAGE_LINE);
@@ -132,10 +133,10 @@ public class ListColoniesCommand extends AbstractSingleCommand
     @NotNull
     @Override
     public List<String> getTabCompletionOptions(
-                                                 @NotNull final MinecraftServer server,
-                                                 @NotNull final ICommandSender sender,
-                                                 @NotNull final String[] args,
-                                                 @Nullable final BlockPos pos)
+            @NotNull final MinecraftServer server,
+            @NotNull final ICommandSender sender,
+            @NotNull final String[] args,
+            @Nullable final BlockPos pos)
     {
         return new ArrayList<>();
     }
