+++ src/main/java/com/minecolonies/coremod/commands/colonycommands/RefreshColonyCommand.java
@@ -1,9 +1,9 @@
-package com.minecolonies.coremod.commands;
+package com.minecolonies.coremod.commands.colonycommands;
 
 import com.minecolonies.coremod.colony.Colony;
 import com.minecolonies.coremod.colony.ColonyManager;
 import com.minecolonies.coremod.colony.IColony;
-import com.minecolonies.api.colony.permissions.Rank;
+import com.minecolonies.coremod.commands.AbstractSingleCommand;
 import com.mojang.authlib.GameProfile;
 import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommandSender;
@@ -18,24 +18,22 @@ import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
 
-import static com.minecolonies.coremod.commands.AbstractSingleCommand.Commands.DELETECOLONY;
-
 /**
  * List all colonies.
  */
-public class DeleteColonyCommand extends AbstractSingleCommand
+public class RefreshColonyCommand extends AbstractSingleCommand
 {
-
-    public static final  String DESC                       = "delete";
+    public static final  String DESC                       = "refresh";
     private static final String NO_COLONY_FOUND_MESSAGE_ID = "Colony with ID %d not found.";
-    private static final String NO_ARGUMENTS               = "Please define a colony to delete";
+    private static final String NO_COLONY_FOUND_MESSAGE    = "Colony with mayor %s not found.";
+    private static final String REFRESH                    = "Refresh succesful!";
 
     /**
      * Initialize this SubCommand with it's parents.
      *
      * @param parents an array of all the parents.
      */
-    public DeleteColonyCommand(@NotNull final String... parents)
+    public RefreshColonyCommand(@NotNull final String... parents)
     {
         super(parents);
     }
@@ -48,54 +46,60 @@ public class DeleteColonyCommand extends AbstractSingleCommand
     }
 
     @Override
-    public boolean canRankUseCommand(@NotNull final Colony colony, @NotNull final EntityPlayer player)
-    {
-        return colony.getPermissions().getRank(player).equals(Rank.OWNER);
-    }
-
-    @Override
     public void execute(@NotNull final MinecraftServer server, @NotNull final ICommandSender sender, @NotNull final String... args) throws CommandException
     {
-        final int colonyId;
+        int colonyId;
+        colonyId = getIthArgument(args, 0, -1);
+        IColony tempColony = ColonyManager.getColony(colonyId);
 
-        if(args.length == 0)
+        if(colonyId == -1 && args.length >= 1)
         {
-            IColony colony = null;
-            if(sender instanceof EntityPlayer)
+            final EntityPlayer player = server.getEntityWorld().getPlayerEntityByName(args[0]);
+            if(player != null)
             {
-                colony = ColonyManager.getIColonyByOwner(((EntityPlayer) sender).worldObj, (EntityPlayer) sender);
+                tempColony = ColonyManager.getIColonyByOwner(server.getEntityWorld(), player);
             }
-
-            if(colony == null)
-            {
-                sender.getCommandSenderEntity().addChatMessage(new TextComponentString(NO_ARGUMENTS));
-                return;
-            }
-            colonyId = colony.getID();
-        }
-        else
-        {
-            colonyId = getIthArgument(args, 0, -1);
-        }
-
-        final Colony colony = ColonyManager.getColony(colonyId);
-        if(colony == null)
-        {
-            sender.getCommandSenderEntity().addChatMessage(new TextComponentString(NO_COLONY_FOUND_MESSAGE_ID));
-            return;
         }
 
         if(sender instanceof EntityPlayer)
         {
+            final UUID mayorID = sender.getCommandSenderEntity().getUniqueID();
+            if (tempColony == null)
+            {
+                tempColony = ColonyManager.getIColonyByOwner(sender.getEntityWorld(), mayorID);
+            }
+
             final EntityPlayer player = (EntityPlayer) sender;
-            if (!canPlayerUseCommand(player, DELETECOLONY, colonyId))
+
+            if (!canPlayerUseCommand(player, Commands.REFRESH_COLONY, colonyId))
             {
                 sender.getCommandSenderEntity().addChatMessage(new TextComponentString(NOT_PERMITTED));
                 return;
             }
         }
 
-        server.addScheduledTask(() -> ColonyManager.deleteColony(colony.getID()));
+        if (tempColony == null)
+        {
+            if (colonyId == -1 && args.length != 0)
+            {
+                sender.addChatMessage(new TextComponentString(String.format(NO_COLONY_FOUND_MESSAGE, args[0])));
+            }
+            else
+            {
+                sender.addChatMessage(new TextComponentString(String.format(NO_COLONY_FOUND_MESSAGE_ID, colonyId)));
+            }
+            return;
+        }
+
+        final Colony colony = ColonyManager.getColony(tempColony.getID());
+        if(colony == null)
+        {
+            sender.getCommandSenderEntity().addChatMessage(new TextComponentString(NO_COLONY_FOUND_MESSAGE_ID));
+            return;
+        }
+
+        sender.getCommandSenderEntity().addChatMessage(new TextComponentString(REFRESH));
+        colony.getPermissions().restoreOwnerIfNull();
     }
 
     @NotNull
