+++ src/main/java/com/minecolonies/coremod/commands/generalcommands/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains all general commands.
+ */
+package com.minecolonies.coremod.commands.generalcommands;
